Move the container_sequence_header logic in EncoderConfig
It does not need any Context information.
